Update _WatchCartoonOnline.py
Updated regex for cizgifilmlerizle. I don't know where else this resolver is used so I created a copy of DoResolve as DoResolve2 and changed that one in case the url is cizgifilmlerizle.
My previous pull was accepted but not released, I guess it's because the version string wasn't incremented. How about merging this and making a new release (e.g. 1.0.6)?
I'll buy you a beer some day. ;)
Cheers